requirements.txt
librosa
pedalboard
pyrubberband
pydub
numpy
soundfile
chromadb
tqdm
pytest
torch
laion-clap
transformers
PyQt6
google-genai
audiocraft
